% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuSection.R
\name{badge}
\alias{badge}
\alias{label}
\title{Create a simple Badge/Label.}
\usage{
badge(label, color)

label(label, status)
}
\arguments{
\item{label}{label name.}

\item{color}{color defined by \link{validColors} (it can be an array).}

\item{status}{status array defined by \link{validStatuses} (it can be an array).}
}
\description{
Dashboard menus can contains badges and labels to help users identify
necessary actions.
}
\examples{
if (interactive()) {
# AdminLTE example
library(adminlte)
shiny::shinyApp(
  ui = dashPage(
    dashHeader(),
    dashSideBar(menuTab(
        menuSection("SECTION  1"),
        menuItem("Dashboard", tabName = "dashboard",
         icon = icon("dashboard"), badges = badge("new", "green")),
        menuItem(text = "Charts", tabName = "charts",
         icon = icon("pie-chart"), status = label("4", "primary"),
            menuSubItem("ChartJS", "charts_1", icon = icon("circle-o")),
            menuSubItem("Morris", "charts_2", icon = icon("circle-o"))
            )
        )),
    dashBody(),
    dashFooter()
  ),
  server = function(input, output) { }
)
}
}
