% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{box}
\alias{box}
\title{Dashboard box container}
\usage{
box(..., title = NULL, footer = NULL, status = NULL,
  solidHeader = FALSE, background = NULL, width = NULL, height = NULL,
  collapsible = FALSE, collapsed = FALSE)
}
\arguments{
\item{...}{1}

\item{title}{1}

\item{footer}{1}

\item{status}{1}

\item{solidHeader}{1}

\item{background}{1}

\item{width}{1}

\item{height}{1}

\item{collapsible}{1}

\item{collapsed}{1}
}
\description{
Dashboard box container
}
