% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{breadcrumb}
\alias{breadcrumb}
\title{Dash board breadcrumbs}
\usage{
breadcrumb(x)
}
\arguments{
\item{x}{An object created by \code{shiny::tags} or a character array.}
}
\description{
Dash board breadcrumbs
}
\examples{

library(adminlte)
breadcrumb(c("Main", "Sub-Page", "Sub-Sub-Page"))

}
