% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashFooter.R
\name{dashFooter}
\alias{dashFooter}
\title{Dashboard footer}
\usage{
dashFooter(...)
}
\arguments{
\item{...}{An object created by \code{shiny::tags}.}
}
\description{
Dashboard footer
}
\examples{

dashFooter(
    shiny::tags$div(
        class = "pull-right hidden-xs",
        shiny::tags$b("Version", "2.3.8")),
        shiny::tags$strong("Copyright ",
        shiny::icon("copyright"),
        " 2014-2016",
        shiny::tags$a(href = "http://almsaeedstudio.com",
        "Almsaeed Studio"),
        "."),"All rights reserved.")

}
