% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashHeader.R
\name{dashHeader}
\alias{dashHeader}
\title{Dashboard Header}
\usage{
dashHeader(logo = dashLogo(), ...)
}
\arguments{
\item{logo}{An object created by \code{\link{dashLogo}}.}

\item{...}{Objects created by \code{menuItem}.}
}
\description{
It contains logos, dropdown menus: messages, notifications, tasks and user menus
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
# Basic dashboard page template
library(shiny)

cheader <- dashHeader()
shinyApp(
  ui = dashboardPage(
    cheader,
    dashSidebar(),
    dashBody(),
    dashFooter()
  ),
  server = function(input, output) { }
)
}
}
