% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashLogo.R
\name{dashLogo}
\alias{dashLogo}
\title{Dashboard Logo}
\usage{
dashLogo(mini = NULL, large = NULL)
}
\arguments{
\item{mini}{An object created by \code{shiny::tags}.}

\item{large}{An object created by \code{shiny::tags}.}
}
\description{
Dashboard Logo
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
# Basic dashboard page template
library(adminlte)
clogo <- dashLogo(mini = shiny::tags$b("ALTE"),
                  large = shiny::tags$b("AdminLTE")
         )
cheader <- dashHeader(clogo)
shinyApp(
  ui = dashPage(
    cheader,
    dashSideBar(),
    dashBody()
  ),
  server = function(input, output) { }
)
}
}
