% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashPage.R
\name{dashPage}
\alias{dashPage}
\title{AdminLTE Dashboard page}
\usage{
dashPage(header, sidebar, body, footer = NULL,
  title = "AdminLTE 2 | General UI", style = "height: auto;",
  skin = "blue")
}
\arguments{
\item{header}{An object created by \code{\link{dashHeader}}.}

\item{sidebar}{An object created by \code{\link{dashSideBar}}.}

\item{body}{An object created by \code{\link{dashBody}}.}

\item{footer}{An object created by \code{\link{dashFooter}}.}

\item{title}{Browser's title bar.}

\item{style}{A css style.}

\item{skin}{A color theme. One of \code{"blue"}, \code{"black"},
\code{"purple"}, \code{"green"}, \code{"red"}, or \code{"yellow"}.}
}
\description{
Complete Dashboard page based on AdminLTE free theme
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
# Basic dashboard page template
library(shiny)
shinyApp(
  ui = dashPage(
    dashHeader(),
    dashSideBar(),
    dashBody(),
    dashFooter()
  ),
  server = function(input, output) { }
)
}

}
