% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashSideBar.R
\name{dashSideBar}
\alias{dashSideBar}
\title{Dashboar sidebar}
\usage{
dashSideBar(..., width = NULL)
}
\arguments{
\item{...}{An object created by \code{shiny::tags}.}

\item{width}{The width of the sidebar in css units.}
}
\description{
Dashboar sidebar
}
\examples{

}
