% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{infoBox}
\alias{infoBox}
\alias{infoBox2}
\title{Dashboard infobox}
\usage{
infoBox(title, value = NULL, subtitle = NULL,
  icon = shiny::icon("bar-chart"), color = "aqua", width = c(3, 3, 6, 6),
  href = NULL, fill = FALSE)

infoBox2(value = NULL, subtitle = NULL, icon = shiny::icon("bar-chart"),
  color = "aqua", width = c(3, 3, 6, 6))
}
\arguments{
\item{title}{An object created by \code{shiny::tags}.}

\item{value}{An number (numeric, integer).}

\item{subtitle}{1}

\item{icon}{1}

\item{color}{1}

\item{width}{1}

\item{href}{1}

\item{fill}{1}
}
\description{
Dashboard infobox
}
