% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuSection.R
\name{menuSection}
\alias{menuSection}
\title{SideBar menu section}
\usage{
menuSection(x)
}
\arguments{
\item{x}{Section name}
}
\description{
SideBar menu section
}
\examples{
if (interactive()) {
# Basic dashboard page template
library(shiny)
shinyApp(
  ui = dashPage(
    dashHeader(),
    dashSideBar(menuTab(
        menuSection("SECTION  1"),
        menuItem("Dashboard", tabName = "dashboard", icon = icon("dashboard"))
        )),
    dashBody(),
    dashFooter()
  ),
  server = function(input, output) { }
)
}
}
