% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuTab.R
\name{menuTab}
\alias{menuTab}
\title{Dashboard sidebar menu container}
\usage{
menuTab(id, ...)
}
\arguments{
\item{id}{menu identification}

\item{...}{menu components}
}
\description{
Dashboard sidebar menu container
}
\examples{


}
