% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{rExample}
\alias{rExample}
\title{adminlte examples}
\usage{
rExample(ex = "dash1")
}
\arguments{
\item{ex}{name of the example

the list of implemented examples is:

\itemize{
  \item \code{dash1}
}}
}
\description{
adminlte examples
}
\examples{
if (interactive()) {
# Basic dashboard page template
library(adminlte)

## call shiny application
runExample()

}
}
