% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashSideBar.R
\name{dashSideBar}
\alias{dashSideBar}
\title{Dashboar sidebar}
\usage{
dashSideBar(..., width = NULL)
}
\arguments{
\item{...}{An object created by \code{shiny::tags}.}

\item{width}{The width of the sidebar in css units.}
}
\description{
Dashboar sidebar
}
\examples{
if (interactive()) {
# AdminLTE example
library(adminlte)
shiny::shinyApp(
  ui = dashPage(
    dashHeader(),
    dashSideBar(menuTab(
        menuSection("SECTION  1"),
        menuItem("Dashboard", tabName = "dashboard", icon = icon("dashboard")),
        menuItem(text = "Charts", tabName = "menu1_item4", icon = icon("pie-chart"),
            menuSubItem("ChartJS", "menu1_item4_subitem1", icon = icon("circle-o")),
            menuSubItem("Morris", "menu1_item4_subitem2", icon = icon("circle-o"))
            )
        )),
    dashBody(),
  ),
  server = function(input, output) { }
)
}
}
