% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuTab.R
\name{menuTab}
\alias{menuTab}
\title{Dashboard sidebar menu container}
\usage{
menuTab(id, ...)
}
\arguments{
\item{id}{menu identification}

\item{...}{menu components}
}
\description{
Dashboard sidebar menu container
}
\examples{
if (interactive()) {
# AdminLTE example
library(adminlte)
shiny::shinyApp(
  ui = dashPage(
    dashHeader(),
    dashSideBar(menuTab(
        menuSection("SECTION  1"),
        menuItem("Dashboard", tabName = "dashboard", icon = icon("dashboard")),
        menuItem(text = "Charts", tabName = "menu1_item4", icon = icon("pie-chart"),
            menuSubItem("ChartJS", "menu1_item4_subitem1", icon = icon("circle-o")),
            menuSubItem("Morris", "menu1_item4_subitem2", icon = icon("circle-o"))
            )
        )),
    dashBody(),
  ),
  server = function(input, output) { }
)
}
}
