% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPanel.R
\name{userPanel}
\alias{userPanel}
\title{Dashboard user panel.}
\usage{
userPanel(user_nm = "User Name", img_src = "avatar.png", ...,
  img_alt = "User Image")
}
\arguments{
\item{user_nm}{User name.}

\item{img_src}{image path.}

\item{...}{An object created by \code{shiny::tag}.}

\item{img_alt}{aimga alternative text.}
}
\description{
Dashboard user panel.
}
\examples{

userPanel(
    user_nm = "User Image",
    img_src = "avatar.png",
    img_alt = "User Image",
    shiny::tags$a(href = "#", shiny::icon("circle text-success"), "Online"))

}
